Refactor: Split employee API into separate files
